+++ go/vt/vtgate/planbuilder/filter.go
@@ -17,11 +17,7 @@ limitations under the License.
 package planbuilder
 
 import (
-	"vitess.io/vitess/go/vt/sqlparser"
-	"vitess.io/vitess/go/vt/vterrors"
 	"vitess.io/vitess/go/vt/vtgate/engine"
-	"vitess.io/vitess/go/vt/vtgate/evalengine"
-	"vitess.io/vitess/go/vt/vtgate/planbuilder/plancontext"
 )
 
 type (
@@ -33,38 +29,6 @@ type (
 )
 
 var _ logicalPlan = (*filter)(nil)
-
-func resolveFromPlan(ctx *plancontext.PlanningContext, plan logicalPlan, canPushProjection bool) evalengine.ColumnResolver {
-	return func(expr *sqlparser.ColName) (int, error) {
-		offset, added, err := pushProjection(ctx, &sqlparser.AliasedExpr{Expr: expr}, plan, true, true, false)
-		if err != nil {
-			return 0, err
-		}
-		if added && !canPushProjection {
-			return 0, vterrors.VT13001("column should not be pushed to projection while doing a column lookup")
-		}
-		return offset, nil
-	}
-}
-
-// newFilter builds a new filter.
-func newFilter(ctx *plancontext.PlanningContext, plan logicalPlan, expr sqlparser.Expr) (*filter, error) {
-	predicate, err := evalengine.Translate(expr, &evalengine.Config{
-		ResolveColumn: resolveFromPlan(ctx, plan, false),
-		ResolveType:   ctx.SemTable.TypeForExpr,
-		Collation:     ctx.SemTable.Collation,
-	})
-	if err != nil {
-		return nil, err
-	}
-	return &filter{
-		logicalPlanCommon: newBuilderCommon(plan),
-		efilter: &engine.Filter{
-			Predicate:    predicate,
-			ASTPredicate: expr,
-		},
-	}, nil
-}
 
 // Primitive implements the logicalPlan interface
 func (l *filter) Primitive() engine.Primitive {
