+++ go/vt/vtgate/planbuilder/projection.go
@@ -22,7 +22,6 @@ import (
 	"vitess.io/vitess/go/vt/sqlparser"
 	"vitess.io/vitess/go/vt/vterrors"
 	"vitess.io/vitess/go/vt/vtgate/engine"
-	"vitess.io/vitess/go/vt/vtgate/evalengine"
 	"vitess.io/vitess/go/vt/vtgate/planbuilder/plancontext"
 	"vitess.io/vitess/go/vt/vtgate/semantics"
 )
@@ -41,28 +40,8 @@ var _ logicalPlan = (*projection)(nil)
 
 // Wireup implements the logicalPlan interface
 func (p *projection) Wireup(ctx *plancontext.PlanningContext) error {
-	if p.primitive != nil {
-		// if primitive is not nil, it means that the horizon planning in the operator phase already
-		// created all the needed evalengine expressions.
-		// we don't need to do anything here, let's just shortcut out of this call
-		return p.source.Wireup(ctx)
-	}
-
-	columns := make([]evalengine.Expr, 0, len(p.columns))
-	for _, expr := range p.columns {
-		convert, err := evalengine.Translate(expr, &evalengine.Config{
-			ResolveColumn: resolveFromPlan(ctx, p.source, false),
-			ResolveType:   ctx.SemTable.TypeForExpr,
-			Collation:     ctx.SemTable.Collation,
-		})
-		if err != nil {
-			return err
-		}
-		columns = append(columns, convert)
-	}
-	p.primitive = &engine.Projection{
-		Cols:  p.columnNames,
-		Exprs: columns,
+	if p.primitive == nil {
+		return vterrors.VT13001("should already be done")
 	}
 
 	return p.source.Wireup(ctx)
@@ -106,22 +85,4 @@ func (p *projection) Primitive() engine.Primitive {
 	}
 	p.primitive.Input = p.source.Primitive()
 	return p.primitive
-}
-
-// addColumn is used to add a column output for the projection.
-// This is the only function that should be used to add  columns to projection
-func (p *projection) addColumn(idx *int, column sqlparser.Expr, columnName string) (int, error) {
-	var offset int
-	if idx == nil {
-		p.unorderedColumnIdx++
-		offset = len(p.columns) - p.unorderedColumnIdx
-	} else {
-		offset = *idx
-	}
-	if p.columnNames[offset] != "" || p.columns[offset] != nil {
-		return -1, vterrors.VT13001("overwriting columns in projection is not permitted")
-	}
-	p.columns[offset] = column
-	p.columnNames[offset] = columnName
-	return offset, nil
 }
