+++ go/vt/vtgate/planbuilder/select.go
@@ -85,6 +85,7 @@ func gen4SelectStmtPlanner(
 
 	if shouldRetryAfterPredicateRewriting(plan) {
 		// by transforming the predicates to CNF, the planner will sometimes find better plans
+		// TODO: this should move to the operator side of planning
 		plan2, tablesUsed := gen4PredicateRewrite(stmt, getPlan)
 		if plan2 != nil {
 			return newPlanResult(plan2.Primitive(), tablesUsed...), nil
@@ -232,8 +233,6 @@ func newBuildSelectPlan(
 		return nil, nil, err
 	}
 
-	optimizePlan(plan)
-
 	if err = plan.Wireup(ctx); err != nil {
 		return nil, nil, err
 	}
@@ -247,108 +246,6 @@ func createSelectOperator(ctx *plancontext.PlanningContext, selStmt sqlparser.Se
 	}
 
 	return operators.PlanQuery(ctx, selStmt)
-}
-
-// optimizePlan removes unnecessary simpleProjections that have been created while planning
-func optimizePlan(plan logicalPlan) {
-	for _, lp := range plan.Inputs() {
-		optimizePlan(lp)
-	}
-
-	this, ok := plan.(*simpleProjection)
-	if !ok {
-		return
-	}
-
-	input, ok := this.input.(*simpleProjection)
-	if !ok {
-		return
-	}
-
-	for i, col := range this.eSimpleProj.Cols {
-		this.eSimpleProj.Cols[i] = input.eSimpleProj.Cols[col]
-	}
-	this.input = input.input
-}
-
-func planLimit(limit *sqlparser.Limit, plan logicalPlan) (logicalPlan, error) {
-	if limit == nil {
-		return plan, nil
-	}
-	rb, ok := plan.(*route)
-	if ok && rb.isSingleShard() {
-		rb.SetLimit(limit)
-		return plan, nil
-	}
-
-	lPlan, err := createLimit(plan, limit)
-	if err != nil {
-		return nil, err
-	}
-
-	// visit does not modify the plan.
-	_, err = visit(lPlan, setUpperLimit)
-	if err != nil {
-		return nil, err
-	}
-	return lPlan, nil
-}
-
-func planHorizon(ctx *plancontext.PlanningContext, plan logicalPlan, in sqlparser.SelectStatement, truncateColumns bool) (logicalPlan, error) {
-	switch node := in.(type) {
-	case *sqlparser.Select:
-		hp := horizonPlanning{
-			sel: node,
-		}
-
-		var err error
-		plan, err = hp.planHorizon(ctx, plan, truncateColumns)
-		if err != nil {
-			return nil, err
-		}
-		plan, err = planLimit(node.Limit, plan)
-		if err != nil {
-			return nil, err
-		}
-	case *sqlparser.Union:
-		var err error
-		rb, isRoute := plan.(*route)
-		if !isRoute && ctx.SemTable.NotSingleRouteErr != nil {
-			return nil, ctx.SemTable.NotSingleRouteErr
-		}
-		if isRoute && rb.isSingleShard() {
-			err = planSingleRoutePlan(node, rb)
-		} else {
-			plan, err = planOrderByOnUnion(ctx, plan, node)
-		}
-		if err != nil {
-			return nil, err
-		}
-
-		plan, err = planLimit(node.Limit, plan)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return plan, nil
-
-}
-
-func planOrderByOnUnion(ctx *plancontext.PlanningContext, plan logicalPlan, union *sqlparser.Union) (logicalPlan, error) {
-	qp, err := operators.CreateQPFromSelectStatement(ctx, union)
-	if err != nil {
-		return nil, err
-	}
-	hp := horizonPlanning{
-		qp: qp,
-	}
-	if len(qp.OrderExprs) > 0 {
-		plan, err = hp.planOrderBy(ctx, qp.OrderExprs, plan)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return plan, nil
 }
 
 func isOnlyDual(sel *sqlparser.Select) bool {
